Carico cartella chess_view
Questa cartella va a sostituire la cartella chessboard
