+++ chess_view/Scacchiera2.java
@@ -1,21 +1,22 @@
-package chessboard;
+package chess_view;
 
 import java.awt.*;
 import java.awt.event.*;
 
 import javax.swing.*;
 
-public class Scacchiera extends JFrame {
+public class Scacchiera2 extends JFrame {
 
 	public static final int WIDTH = 700;
 	public static final int HEIGHT = 700;
 	
 	public static void main(String[] args) {
 		
-		new Scacchiera().setVisible(true);
+		new Scacchiera2().setVisible(true);
+		new OptionWindow().setVisible(true);
 	}
 	
-	public Scacchiera() {
+	public Scacchiera2() {
 		
 		super("Scacchiera");
 		setSize(WIDTH, HEIGHT);
@@ -26,7 +27,8 @@ public class Scacchiera extends JFrame {
 		addWindowListener(new WindowAdapter() {
 			@Override
 			public void windowClosing(WindowEvent e) {
-				new ConfirmWindow().setVisible(true);
+				ConfirmWindow thirdWindow = new ConfirmWindow();
+				thirdWindow.setVisible(true);
 			}
 		});
 		setLayout(new BorderLayout());
@@ -51,12 +53,12 @@ public class Scacchiera extends JFrame {
 		
 		add(numbersPanel(), BorderLayout.EAST);
 		
-		add(northPanel(), BorderLayout.NORTH);
-		
-		add(southPanel(), BorderLayout.SOUTH);
+		add(lettersPanel(), BorderLayout.NORTH);
+	
+		add(lettersPanel(), BorderLayout.SOUTH);
 				
 	}
-
+	
 	private JButton a8_g8_a2_g2_Grid(int i, int j) {
 		
 		JButton b = new JButton();
@@ -172,39 +174,5 @@ public class Scacchiera extends JFrame {
 		
 		return lettersPanel;
 	}
-	
-	private JPanel northPanel() {
-		
-		JPanel northPanel = new JPanel();
-		northPanel.setLayout(new BorderLayout());
-		
-		JPanel northLettersPanel = lettersPanel();
-		northPanel.add(northLettersPanel, BorderLayout.SOUTH);
-		
-		// DA COMPLETARE
-		
-		return northPanel;
-	}
-	
-	private JPanel southPanel() {
-		
-		JPanel southPanel = new JPanel();
-		southPanel.setLayout(new BorderLayout());
-		
-		JPanel southLettersPanel = lettersPanel();
-		southPanel.add(southLettersPanel, BorderLayout.NORTH);
-		
-		JTextField textFieldLeft = new JTextField();
-		southPanel.add(textFieldLeft, BorderLayout.WEST);
-		textFieldLeft.setPreferredSize(new Dimension(100, 0));
-		JTextField textFieldCenter = new JTextField();
-		southPanel.add(textFieldCenter, BorderLayout.CENTER);
-		JTextField textFieldRight = new JTextField();
-		textFieldRight.setPreferredSize(new Dimension(100, 0));
-		southPanel.add(textFieldRight, BorderLayout.EAST);
-		
-		return southPanel;
-	}
 
 }
-
