+++ chess_view/images/pieces/black_bishop.png
