+++ chess_view/images/pieces/black_king.png
