+++ chess_view/images/pieces/black_pawn.png
