+++ chess_view/images/pieces/black_queen.png
