+++ chess_view/images/pieces/black_rook.png
