+++ chess_view/images/pieces/white_bishop.png
