+++ chess_view/images/pieces/white_king.png
