+++ chess_view/images/pieces/white_knight.png
