+++ chess_view/images/pieces/white_pawn.png
