+++ chess_view/images/pieces/white_queen.png
