+++ chess_view/images/pieces/white_rook.png
